loading .obj models

+++ asset/include/asset/mesh.h
@@ -26,15 +26,15 @@ public:
   virtual ~Mesh();
   const std::vector<Vertex>& Vertices() const { return vertices_; }
   Mesh& Vertices(const std::vector<Vertex>& vertices) { vertices_ = vertices; return *this; }
-  const std::vector<uint16_t>& Indices() const { return indices_; }
-  Mesh& Indices(const std::vector<uint16_t>& indices) { indices_ = indices; return *this; }
+  const std::vector<uint32_t>& Indices() const { return indices_; }
+  Mesh& Indices(const std::vector<uint32_t>& indices) { indices_ = indices; return *this; }
 public: // Serialization
   static const AssetId assetId;
   static size_t Serialize(Asset* asset, std::ostream& stream, bool binary);
   static MeshPtr Deserialize(const std::string& name, std::istream& stream, size_t size, bool binary);
 private:
   std::vector<Vertex> vertices_;
-  std::vector<uint16_t> indices_;
+  std::vector<uint32_t> indices_;
 };
 
 } // !namespace asset
